+++ CHANGELOG.md
@@ -5,6 +5,11 @@ is based on https://keepachangelog.com/en/1.1.0/.
 
 ## Unreleased
 
+### Added
+
+- A fn `tracer::Tracer::qual_status` for retrieving the qualification status
+  after a trace has ended or was lost.
+
 ## 0.8.0 - 2026-01-31
 
 ### Added